==> checkpoints-tasks/checkpoint-4/features_new/Area_parsing_new/01 Selenium_download_reestrs.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.by import By
from tqdm import tqdm_notebook as tqdmn
import pandas as pd
from time import sleep

#В этой части собираем ссылки на все реестры домов по всем регионам с сайта www.reformagkh.ru

#Загружаем драйвер. 

print("Loading driver...") 
option = webdriver.ChromeOptions() #опции по умолчанию
driver = webdriver.Chrome("./chromedriver", options=option)
driver.get("about:blank")

#собираем ссылочки в словарик

n = 1
d = {}
while True:
    driver.get(f"https://www.reformagkh.ru/opendata?cids=house_management&page={n}&pageSize=60")
    cnt = driver.find_elements(By.CSS_SELECTOR, 'p.lh-27.f-18.fw-700')
    lnk = driver.find_elements(By.CSS_SELECTOR, 'div.my-3 > a')
    if not cnt:
        break
    for i, c in enumerate(cnt):
        l = lnk[i]
        d[c.text] = l.get_attribute('href')
    n += 1
d

#Создаём список со ссылками. Выводим на экран, чтобы посмотреть на них

urls = []
for elem in d:
    if elem.startswith('Реестр домов по '):
        print(elem,":")
        url = d[elem]
        print(url)
        urls.append(url)

#Проходим по ссылочкам, скачиваем автоматически. Спим 1,5 секунды, чтобы успело скачаться.

for u in tqdmn(urls, leave=False):
    driver.get(u)
    sleep(1.5)

driver.close()

driver = None
