==> checkpoints-tasks/checkpoint-4/features_new/Area_parsing_new/02 Selenium_create_csv.ipynb <==
import pandas as pd
from os import walk
from os.path import join as pjoin
from tqdm import tqdm

#В этой части вытаскиваем нужные данные из скаченных файлов и собираем в csv

#Скачанные файлы заранее положили в отдельную папку "reestrs"

PATH = "reestrs"
files = [(pjoin(PATH, f), f) for f in list(walk(PATH))[0][2] if f.endswith(".csv")]
#создаём список из названий файлов - получаем список списков. 
#Выбираем результат [0][2] - получаем список файлов. И джойним каждый файл с названием папки

len(files) #Ничего не потерялось

cols = ['id', 'address', 'area_residential'] 
#нужные нам колонки из файлов с данными

values = []

for filepath, filename in tqdm(files):
    df = pd.read_csv(filepath, delimiter=";", low_memory=False)[cols].dropna()
    for _, i in df.iterrows():
        value = []
        value.extend([i[col] for col in cols])
        value.append(filename)
        values.append(value)

cols.append('filename')
df_united = pd.DataFrame(values, columns=cols)
df_united

df_united.to_csv('united_reestr.csv')
