==> checkpoints-tasks/checkpoint-4/features_new/Area_parsing_new/04 Selenium_url_to_coordinates.ipynb <==
import pandas as pd

#Парсим url адреса и создаём в csv с адресами новые столбцы с lat и lon

df = pd.read_csv("nts.csv")
df.head()

FAILED_URLS = []
def crds(o):
    #Парсим url-адрес и достаём координаты
    try:
        return [float(i) for i in o.split('&zoom')[0].split("?center=")[1].split("%2C")]
    except:
        FAILED_URLS.append(o)
        return None

FAILED_URLS = []
df['url'] = df['url'].apply(crds)
set(FAILED_URLS)

#Смотрим адреса, для которых не удалось найти координаты
df_err = df[df['url'].isna()]
df_all = pd.read_csv('united_reestr.csv')[['id', 'address']].drop_duplicates()
df_postparse = pd.merge(df_all, df_err, left_on="id", right_on="id", how='inner')
df_postparse

df.dropna()

#Удаляем адреса, для которых не удалось найти координаты (позже можно их найти вручную)

df_correct = df[~df['url'].isna()]
df_all2 = pd.read_csv('united_reestr.csv').drop_duplicates()
df_all_correct = pd.merge(df_all2, df_correct, left_on="id", right_on="id", how='inner')
df_all_correct = df_all_correct.drop('Unnamed: 0', axis='columns').rename(columns={'url': 'coords'})
df_all_correct

#Функция разбивает координаты на широту и долготу
def correct_coords(row, side):
    return row['coords'][side]
df_all_correct['lat'] = df_all_correct.apply(correct_coords, args=(0,), axis=1)
df_all_correct['lon'] = df_all_correct.apply(correct_coords, args=(1,), axis=1)
df_final = df_all_correct.drop('coords', axis='columns')
df_final

df_final.to_csv("final_csv.csv", index=False)

df_final_test = pd.read_csv("final_csv.csv")
df_final_test.head(-5)
